Add logs to Post(), fix int formatting in a log, and small doc fix
- Remove a sentence from the config doc about cookies which
  was no longer accurate
- Add log statements to Post() to help the user debug proxy
  vs no proxy setting
- In a log statement that used to say `RefreshToken: Received
  (non-OK) status 400.000000 from identity provider when
  refreshing the access token.` make the `400.000000` just
  say `400`

+++ src/filters/oidc/oidc_filter.cc
@@ -451,8 +451,7 @@ std::shared_ptr<TokenResponse> OidcFilter::RefreshToken(
   http::status status = retrieved_token_response->result();
   if (status != boost::beast::http::status::ok) {
     spdlog::warn("{}: Received (non-OK) status {} from identity provider when refreshing the access token.", __func__,
-                 std::to_string(
-                     static_cast<double>(status)));
+                 std::to_string(static_cast<unsigned>(status)));
     return nullptr;
   }
